requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
python-dotenv~=1.0.1
discord.py~=2.3.2
mysql-connector-python~=8.3.0